Extract CityGML surface reader
