+++ src/PlateauResoniteLink/Application/Importing/CityGmlCoordinateTextParser.cs
@@ -1,0 +1,40 @@
+using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+
+namespace PlateauResoniteLink.Application.Importing;
+
+internal static class CityGmlCoordinateTextParser
+{
+    internal static double[] ParseDoubles(string value)
+    {
+        return value
+            .Split((char[]?)null, StringSplitOptions.RemoveEmptyEntries)
+            .Select(token => double.Parse(token, CultureInfo.InvariantCulture))
+            .ToArray();
+    }
+
+    internal static List<Float2> ParseTextureCoordinates(string value)
+    {
+        double[] ordinates = ParseDoubles(value);
+        List<Float2> coordinates = [];
+        for (int index = 0; index + 1 < ordinates.Length; index += 2)
+        {
+            coordinates.Add(new Float2(ordinates[index], ordinates[index + 1]));
+        }
+
+        if (coordinates.Count > 1 && AreSameUV(coordinates[0], coordinates[^1]))
+        {
+            coordinates.RemoveAt(coordinates.Count - 1);
+        }
+
+        return coordinates;
+    }
+
+    private static bool AreSameUV(Float2 left, Float2 right)
+    {
+        return Math.Abs(left.X - right.X) < 1e-8
+            && Math.Abs(left.Y - right.Y) < 1e-8;
+    }
+}
